update GdbController API, add IoMangager, other dev improvements
