+++ docs/api/iomanager.md
@@ -1,0 +1,1 @@
+::: pygdbmi.IoManager
