+++ pygdbmi/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "0.9.0.3"
+__version__ = "0.10.0.0b0"
